requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.1.13
djongo==1.3.6
dnspython==2.6.1
google-auth==2.30.0
idna==3.7
logger==1.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pymongo==3.12.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
sqlparse==0.2.4
tzdata==2024.1
urllib3==2.2.1
logging== 0.4.9.6
